refactor(card libs ui): renamed small-info-card folder to card folder
