+++ libs/ui/src/components/card/SmallInfoCard.spec.tsx
