+++ libs/ui/src/index.ts
@@ -1,5 +1,5 @@
 export * from './components/icon-button-styled/IconButtonStyled';
-export * from './components/small-info-card/SmallInfoCard';
+export * from './components/card/SmallInfoCard';
 export * from './context/theme-context/ThemeContext';
 export * from './components/error-message/ErrorMessage';
 export * from './components/fields/input-text-field/InputTextField';
